+++ single_kernel_postgresql/charms/k8s_charm.py
@@ -26,6 +26,7 @@ class PostgreSQLK8sCharm(AbstractPostgreSQLCharm):
             "Workload must be an instance of K8sWorkload"
         )
         self.k8s_manager = K8sManager(self.state, self.workload)
+        self.state.resource_provider = self.k8s_manager
 
     @property
     def postgresql(self) -> PostgreSQL:
@@ -54,8 +55,6 @@ class PostgreSQLK8sCharm(AbstractPostgreSQLCharm):
         return K8sWorkload(
             charm_dir=self.charm_dir,
             container=self.unit.get_container(CONTAINER_NAME),
-            unit_name=self.unit.name,
-            namespace=self.model.name,
         )
 
     @property
