+++ single_kernel_postgresql/workload/k8s.py
@@ -12,24 +12,13 @@ from types import SimpleNamespace
 
 from charmlibs import pathops
 from charmlibs.pathops import PathProtocol
-from lightkube import Client
-from lightkube.core.exceptions import ApiError
-from lightkube.resources.core_v1 import Node, Pod
 from ops import Container, ModelError
 from ops.pebble import Plan, ServiceStatus
 
-from single_kernel_postgresql.config.exceptions import (
-    DeployedWithoutTrustError,
-    PostgreSQLFileOperationError,
-)
+from single_kernel_postgresql.config.exceptions import PostgreSQLFileOperationError
 from single_kernel_postgresql.config.literals import (
     DIR_PERMISSIONS_READONLY,
     K8S_POSTGRESQL_SERVICE_NAME,
-)
-from single_kernel_postgresql.utils import (
-    any_cpu_to_cores,
-    any_memory_to_bytes,
-    unit_name_to_pod_name,
 )
 from single_kernel_postgresql.workload.base import BaseWorkload
 from single_kernel_postgresql.workload.paths.base import Paths as BasePaths
@@ -41,20 +30,15 @@ logger = logging.getLogger(__name__)
 class K8sWorkload(BaseWorkload):
     """Kubernetes PostgreSQL Workload."""
 
-    def __init__(self, charm_dir: Path, container: Container, *, unit_name: str, namespace: str):
+    def __init__(self, charm_dir: Path, container: Container):
         """Initialize workload.
 
         Args:
             charm_dir: the path to charm code.
             container: the Container instance.
-            unit_name: the Juju unit name (e.g. "postgresql-k8s/0"), used to resolve
-                this unit's own pod for K8s resource-limit lookups.
-            namespace: the Juju model name, i.e. the K8s namespace the pod lives in.
         """
         super().__init__(charm_dir=charm_dir)
         self.container = container
-        self.unit_name = unit_name
-        self.namespace = namespace
         self._paths: BasePaths | None = None
 
     def install(self) -> None:
@@ -237,74 +221,3 @@ class K8sWorkload(BaseWorkload):
             return container.can_connect()
         except (RuntimeError, ModelError):
             return False
-
-    def get_available_memory(self) -> int:
-        """Returns the system available memory in bytes."""
-        raise NotImplementedError
-
-    def _get_node_name_for_pod(self) -> str:
-        """Return the node name for this unit's own pod."""
-        client = Client()
-        pod = client.get(Pod, name=unit_name_to_pod_name(self.unit_name), namespace=self.namespace)
-        if pod.spec and pod.spec.nodeName:
-            return pod.spec.nodeName
-        raise RuntimeError("Pod doesn't exist")
-
-    def get_resources_limits(self, container_name: str) -> dict:
-        """Return resources limits for a given container.
-
-        Args:
-            container_name: name of the container to get resources limits for.
-        """
-        client = Client()
-        pod = client.get(Pod, name=unit_name_to_pod_name(self.unit_name), namespace=self.namespace)
-        if pod.spec:
-            for container in pod.spec.containers:
-                if container.name == container_name and container.resources:
-                    return container.resources.limits or {}
-        return {}
-
-    def get_node_allocable_memory(self) -> int:
-        """Return the allocable memory in bytes for the current K8s node."""
-        client = Client()
-        node = client.get(Node, name=self._get_node_name_for_pod(), namespace=self.namespace)  # type: ignore
-        return any_memory_to_bytes(node.status.allocatable["memory"])
-
-    def get_node_cpu_cores(self) -> int:
-        """Return the number of CPU cores for the current K8s node."""
-        client = Client()
-        node = client.get(Node, name=self._get_node_name_for_pod(), namespace=self.namespace)  # type: ignore
-        return any_cpu_to_cores(node.status.allocatable["cpu"])
-
-    def get_available_resources(self) -> tuple[int, int]:
-        """Returns the available (cpu_cores, memory_bytes) for the workload.
-
-        Raises:
-            DeployedWithoutTrustError: if the K8s API denies access (403), meaning the
-                app wasn't deployed with ``--trust``.
-        """
-        try:
-            cpu_cores = self.get_node_cpu_cores()
-            allocable_memory = self.get_node_allocable_memory()
-            container_limits = self.get_resources_limits(
-                container_name=K8S_POSTGRESQL_SERVICE_NAME
-            )
-        except ApiError as e:
-            if e.status.code == 403:
-                raise DeployedWithoutTrustError from e
-            raise
-
-        if "cpu" in container_limits:
-            cpu_str = container_limits["cpu"]
-            constrained_cpu = any_cpu_to_cores(cpu_str)
-            if constrained_cpu < cpu_cores:
-                logger.debug(f"CPU constrained to {cpu_str} cores from resource limit")
-                cpu_cores = constrained_cpu
-        if "memory" in container_limits:
-            memory_str = container_limits["memory"]
-            constrained_memory = any_memory_to_bytes(memory_str)
-            if constrained_memory < allocable_memory:
-                logger.debug(f"Memory constrained to {memory_str} from resource limit")
-                allocable_memory = constrained_memory
-
-        return cpu_cores, allocable_memory
