+++ tests/unit/test_config_manager.py
@@ -14,11 +14,7 @@ from single_kernel_postgresql.workload.vm import VMWorkload
 def config(substrate):
     mock_charm = Mock()
     mock_container = Mock()
-    workload = (
-        VMWorkload(".")
-        if substrate == Substrates.VM
-        else K8sWorkload(".", mock_container, unit_name="postgresql-k8s/0", namespace="test-model")
-    )
+    workload = VMWorkload(".") if substrate == Substrates.VM else K8sWorkload(".", mock_container)
     with patch(
         "single_kernel_postgresql.workload.base.BaseWorkload.get_postgresql_version",
         return_value="16.6",
